pyproject.toml
[tool.poetry]
name = "mplug"
version = "0.1.0"
description = "A plugin manager for mpv"
authors = ["Michael F. Schönitzer <michael@schoenitzer.de>"]
license = "MPL-2.0"
readme = "README.md"
homepage = "https://github.com/nudin/mplug"
repository = "https://github.com/nudin/mplug"
keywords = ["mpv"]

[tool.poetry.dependencies]
python = "^3.6"
GitPython = "^3.1.7"
requests = "^2.24.0"
importlib_metadata = {version = "^1.7.0", python = "<3.8"}

[tool.poetry.dev-dependencies]
pytest = "^6.0.1"
pytest-mock = "^3.3.0"
coverage = {extras = ["toml"], version = "^5.2.1"}
pytest-cov = "^2.10.1"
mypy = "^0.782"
pylint = "^2.6.0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"

[tool.poetry.scripts]
mplug = "mplug:run"
